+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".EventActivity">
+        <activity android:name=".SearchEventActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
