+++ app/src/main/java/com/segmentify/segmentifysdk/EventActivity.java
@@ -2,11 +2,20 @@ package com.segmentify.segmentifysdk;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
 
+import com.google.firebase.iid.FirebaseInstanceId;
 import com.segmentify.segmentifyandroidsdk.SegmentifyManager;
-import com.segmentify.segmentifyandroidsdk.model.SearchPageModel;
-import com.segmentify.segmentifyandroidsdk.model.SearchResponseModel;
+import com.segmentify.segmentifyandroidsdk.model.CheckoutModel;
+import com.segmentify.segmentifyandroidsdk.model.NotificationModel;
+import com.segmentify.segmentifyandroidsdk.model.NotificationType;
+import com.segmentify.segmentifyandroidsdk.model.PageModel;
+import com.segmentify.segmentifyandroidsdk.model.ProductModel;
+import com.segmentify.segmentifyandroidsdk.model.RecommendationModel;
 import com.segmentify.segmentifyandroidsdk.utils.SegmentifyCallback;
+
+import java.util.ArrayList;
 
 public class EventActivity extends AppCompatActivity {
 
@@ -15,21 +24,7 @@ public class EventActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        SearchPageModel searchPage = new SearchPageModel();
-        searchPage.setQuery("");
-        searchPage.setLang("EN");
-        SegmentifyManager.INSTANCE.sendSearchPageView(searchPage, new SegmentifyCallback<SearchResponseModel>() {
-                    @Override
-                    public void onDataLoaded(SearchResponseModel data) {
-                        if(data!=null){
-                            System.out.println(data);
-                        }
-                    }
-                });
-
-        SegmentifyManager.INSTANCE.sendSearchClickView("product","21046");
-
-        /*PageModel model = new PageModel();
+        PageModel model = new PageModel();
         model.setCategory("Home Page");
 
 
@@ -78,7 +73,6 @@ public class EventActivity extends AppCompatActivity {
 
             }
         });
-*/
 
     }
 }
