+++ app/src/main/java/com/segmentify/segmentifysdk/MyApp.java
@@ -8,7 +8,7 @@ public class MyApp extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
-        SegmentifyManager.INSTANCE.config(this,"8f6b9ae8-7d0e-455d-be6d-bdf7b74efcf7","https://gandalf-dev.segmentify.com","ihalilaltun-dev.me");
+        SegmentifyManager.INSTANCE.config(this,"277ae43c-1ed5-4d64-922d-b4e704de5d2e","https://gandalf-dev.segmentify.com","ihalilaltun-dev.me");
         SegmentifyManager.INSTANCE.setSessionKeepSecond(100000);
         SegmentifyManager.INSTANCE.logStatus(true);
     }
